Collapse benchmarks/src into a single script
Inline `Models.jl` and `DynamicPPLBenchmarks.jl` into `benchmarks.jl` and
convert `benchmarks/Project.toml` from a package to a flat environment,
mirroring Mooncake.jl's `bench/run_benchmarks.jl` layout.

Also: take dim from `length(r.params)` (run_ad already constructed the LDF)
so models are no longer evaluated twice on the success path; switch results
to NamedTuples so `print_results` reads `r.name`/`r.dim`/...; extract
`transform_strategy(islinked)` helper; drop unused imports.
